Rename EmailField components
